chore(refactor): moved zip logic out of code into cloud build yaml
